% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bare_combine.R
\name{bare_combine}
\alias{bare_combine}
\title{Split a string into a vector of strngs}
\usage{
bare_combine(string)
}
\arguments{
\item{string}{A comma separated string}
}
\value{
A vector of strings based on comma locations
}
\description{
Split a string into a vector of strngs
}
\examples{
bare_combine("t1, t2, t3, t4")
}
