% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_share.R
\name{calculate_share}
\alias{calculate_share}
\title{Calculate share of search}
\usage{
calculate_share(tbl, ma, pal)
}
\arguments{
\item{tbl}{Data frame output from Google Trends API}

\item{ma}{Moving average to apply via Shiny sliderInput}

\item{pal}{Color palette selected by user in Shiny}
}
\value{
list containing 3 plots (share line, share area, zscore area) and 3 tables of data used
}
\description{
Calculates the share of search for each search term and date period.
Share is calculated for raw hits data as well as several moving averages (via Shiny sliderInput). Z-scores are also calculated for each time period and Keyword.
}
\examples{
\dontrun{
calculate_share(tbl)}
}
