% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_search.R
\name{cat_search}
\alias{cat_search}
\title{Search for Google Trend categories by name}
\usage{
cat_search(term)
}
\arguments{
\item{term}{Term to use in the search}
}
\value{
A two column data frame of matched terms
}
\description{
Using str_detect to search, returns partial matches; capitalization doesn't matter
}
\examples{
cat_search("audio")
}
