% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{Google Trends categories}
\format{
A dataframe with 1426 categories and IDs

\describe{
\item{name}{Category name, e.g. Arts & Entertainment}
\item{id}{Numeric ID, e.g., 3}
}
}
\usage{
categories
}
\description{
The category names and IDs used when specifying a category in a trends search.
}
\keyword{datasets}
