% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{data_process}
\alias{data_process}
\title{Clean data for saving}
\usage{
data_process(tbl, calc_tbl, corr_tbl, arg_list)
}
\arguments{
\item{tbl}{Data frame returned by Google Trends API}

\item{calc_tbl}{Calculated table in wide form}

\item{corr_tbl}{Correlation table}

\item{arg_list}{Additional arguments passed through by user}
}
\value{
Excel sheets saved to data subfolder
}
\description{
Cleaning data and compiling into a single Excel sheet
}
\examples{
\dontrun{
data_process(tbl, calc_tbl, corr_tbl, arg_list)}
}
