% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp}
\alias{decomp}
\title{Time Series Decompositions}
\usage{
decomp(tbl, directory, project)
}
\arguments{
\item{tbl}{Data frame returned by Google Trends API}

\item{directory}{Location where decompositions are saved; user will have previously defined directories}

\item{project}{Name of project; previously provided by user}
}
\value{
Data frame and graphs. Data frame with decomp metrics. Graphs are saved into a \verb{/graphs/decompositions/} subdirectory.
}
\description{
If the trends data is longer than 3 years in length, function runs an STL decomposition on each series.

See Hyndman "Forecasting: Principles and Practice" \url{https://otexts.com/fpp3/stlfeatures.html} for details.
}
\examples{
\dontrun{
decomp(tbl, directory, project)}
}
