% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_create.R
\name{directory_create}
\alias{directory_create}
\title{Create directories and sub-directories}
\usage{
directory_create(projname, main_dir, arg_list, permiss, data)
}
\arguments{
\item{projname}{Main folder; defined by user input}

\item{main_dir}{1st level folder where data is stored}

\item{arg_list}{Inputs from the user in list form}

\item{permiss}{Where to set up directory}

\item{data}{Data frame returned from Google Trends API}
}
\value{
NULL folders are set up in specific location
}
\description{
Create directories and sub-directories
}
\examples{
\dontrun{
directory_create(main_dir, arg_list, permiss, data)}
}
