% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find100.R
\name{find100}
\alias{find100}
\title{Find the keyword with peak search interest}
\usage{
find100(tbl)
}
\arguments{
\item{tbl}{A data frame of data returned by Google Trends API}
}
\value{
One row data frame of Trends data with hits score of 100
}
\description{
Find the keyword with peak search interest
}
\examples{
\dontrun{
find100(data)}
}
