% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_trends}
\alias{get_trends}
\title{Query Google Trends API}
\usage{
get_trends(arg_list)
}
\arguments{
\item{arg_list}{list of arguments to use in API call; arguments provided by user through the Shiny app and wrapped into a list}
}
\value{
Data frame returned from Google Trends API
}
\description{
Convenience wrapper for the \code{\link[gtrendsR]{gtrends}} function.

If there are more than 5 terms to query, function will ensure that the most popular term is in each query.
}
\examples{
\dontrun{
get_trends(arg_list)}
}
