% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_corr.R
\name{gtr_corr}
\alias{gtr_corr}
\title{Estimate correlations among Google Trend keywords}
\usage{
gtr_corr(tbl, directory)
}
\arguments{
\item{tbl}{Data frame output from Google Trends API}

\item{directory}{Pre-specified directory of where to save the plot and data.}
}
\value{
Data frame of correlation data
}
\description{
Estimates the correlations across keywords and creates a ggplot correlation heatmap and a correlation table.
The correlation table is put into the global environment for use by those that don't want to save materials to a directory.
}
\examples{
\dontrun{
get_corr(tbl, dir)}
}
