% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_over_time.R
\name{interest_over_time}
\alias{interest_over_time}
\title{Total function to call Google Trends Interest Over Time}
\usage{
interest_over_time()
}
\value{
Nothing is returned. Graphs and Excel workbooks are saved to directory locations.
}
\description{
User is put through a flow of questions. All answers are saved to a list of arguments that are then used to query the Google Trends API.

Function automatically selects the working directory as the location to save all data and graphs, but user has option to change directory location.
}
\examples{
\dontrun{
interest_over_time()}
}
