% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_subtitle.R
\name{ma_subtitle}
\alias{ma_subtitle}
\title{Determine length of Moving Average}
\usage{
ma_subtitle(tbl, agg)
}
\arguments{
\item{tbl}{Data frame of Google Trends data from the API}

\item{agg}{The lengths of the moving averages applied to the data}
}
\value{
String detailing the length of the moving average
}
\description{
Calculate the duration of several moving averages that are applied to data returned from Google Trends API.

Trends data uses different time windows depending on the duration searched. This function determines the length and determines the right unit of time.
}
\examples{
\dontrun{
ma_subtitle(tbl, "ma7")
}
}
