% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ind_plot.R
\name{make_ind_plot}
\alias{make_ind_plot}
\title{Plots for individual keywords}
\usage{
make_ind_plot(tbl, args, directory)
}
\arguments{
\item{tbl}{Data frame returned by Google Trends API}

\item{args}{Various arguments compiled from user input}

\item{directory}{Location to save graphs, provided by user}
}
\value{
png files saved to specified directory
}
\description{
Plots for each keyword; raw data with a hard-coded MA5
}
\examples{
\dontrun{
make_ind_plot(tbl, args, directory)}
}
