% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make_plots}
\alias{make_plots}
\title{Graph the Google Trends data}
\usage{
make_plots(tbl, var, directory, args)
}
\arguments{
\item{tbl}{Data frame returned by Google Trends API}

\item{var}{Variable of choice for graphing}

\item{directory}{Where to save the graphs}

\item{args}{Additional data for captions and labels}
}
\value{
png files saved to graph subfolders
}
\description{
Multiple graphs with all keywords included in the data
}
\examples{
\dontrun{
make_plots(tbl, var, directory, args)}
}
