% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metros}
\alias{metros}
\title{Google Trends US metro codes}
\format{
A data frame with 210 metros by name, id, state code and geo code

\describe{
\item{name}{Name of DMA, e.g., Anchorage AK}
\item{id}{Three digit id code, e.g., 743}
\item{code}{State abbreviation, e.g., AK}
\item{geo}{Geo region code used by Google Trends, e.g., US-AK-74}3
}
}
\usage{
metros
}
\description{
The metro areas and geo codes used by Google in its trends searches. Data was
created from geo code from
\url{"https://trends.google.com/trends/api/explore/pickers/geo"}
}
\keyword{datasets}
