% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_gtr.R
\name{pull_gtr}
\alias{pull_gtr}
\title{Query Google Trends API}
\usage{
pull_gtr(arg_list)
}
\arguments{
\item{arg_list}{list of arguments to use in API call; arguments provided by user through the interactive interface and automatically wrapped into a list}
}
\value{
Data frame returned from Google Trends API
}
\description{
Convenience wrapper for the \code{\link[gtrendsR]{gtrends}} function.

If there are more than 5 terms to query, function will ensure that the most popular term is in each query.

\strong{Note:} Max of 9 terms; geo is hard coded to the US for now
}
\examples{
\dontrun{
pull_gtr(arg_list)}
}
