% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_summary.R
\name{query_summary}
\alias{query_summary}
\title{Summarizing Raw Google Trends Data}
\usage{
query_summary(tbl)
}
\arguments{
\item{tbl}{Raw data pulled via the gtrendsR package}
}
\value{
list with two tables (wide summary and summary stats) and basic line plot
}
\description{
Summarizing Raw Google Trends Data
}
\examples{
\dontrun{
query_summary(tbl)
}
}
