% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_calc.R
\name{share_calc}
\alias{share_calc}
\title{Calculate share of search}
\usage{
share_calc(tbl)
}
\arguments{
\item{tbl}{Data frame output from Google Trends API}
}
\value{
A tidy data frame with share calculations for each term overall, MA3, MA5, MA7, and MA13.
}
\description{
Calculates the share of search for each search term and date period.
Share is calculated for raw hits as well as multiple moving averages.
Z-scores are also calculated for each time period and Keyword.
}
\examples{
\dontrun{
share_calc(tbl)}
}
