% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_keyword_smooth.R
\name{single_keyword_smooth}
\alias{single_keyword_smooth}
\title{Single term smoothed}
\usage{
single_keyword_smooth(tbl, term, ma)
}
\arguments{
\item{tbl}{Data frame with Google Trends data}

\item{term}{Term to isolate}

\item{ma}{Level of moving average}
}
\value{
graph with raw hits value and overlayed moving average line
}
\description{
Single term smoothed
}
\examples{
\dontrun{
single_keyword_smooth(tbl, "cats", ma = 7)
}
}
