% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{States to narrow trends search}
\format{
Dataframe with 50 states and United States

\describe{
\item{name}{full name}
\item{abb}{state abbreviation}
\item{id}{formatted for use}
}
}
\usage{
states
}
\description{
State abbreviations to narrow down the search
}
\keyword{datasets}
