% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xf.R
\name{theme_xf}
\alias{theme_xf}
\title{Theme for ggplot}
\usage{
theme_xf(
  base_family = "Open Sans",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 16,
  plot_title_face = "plain",
  plot_title_margin = 4,
  subtitle_family = "Open Sans",
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = "Open Sans",
  caption_size = 9,
  caption_face = "italic",
  caption_margin = 10,
  axis_title_family = base_family,
  axis_title_size = 13,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(10, 10, 10, 10),
  panel_spacing = ggplot2::unit(0.5, "lines"),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Font family}

\item{base_size}{base size}

\item{plot_title_family}{title font family}

\item{plot_title_size}{title size}

\item{plot_title_face}{title face}

\item{plot_title_margin}{title margin}

\item{subtitle_family}{subtitle family}

\item{subtitle_size}{subtitle size}

\item{subtitle_face}{subtitle face}

\item{subtitle_margin}{subtitle margin}

\item{strip_text_family}{strip text family}

\item{strip_text_size}{strip text font size}

\item{strip_text_face}{strip text font face}

\item{caption_family}{family}

\item{caption_size}{size}

\item{caption_face}{face}

\item{caption_margin}{margin}

\item{axis_title_family}{family}

\item{axis_title_size}{size}

\item{axis_title_face}{face}

\item{axis_title_just}{justification}

\item{plot_margin}{margin}

\item{panel_spacing}{spacing}

\item{grid_col}{grid}

\item{grid}{grid}

\item{axis_col}{axis}

\item{axis}{axis}

\item{ticks}{ticks}
}
\value{
re-themed ggplot graph
}
\description{
Theme for ggplot
}
\examples{
\dontrun{
ggplot(mtcars, aes(x = mpg, y = hp)) +
geom_point() +
theme_xf
}
}
