% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_correlation.R
\name{trend_correlation}
\alias{trend_correlation}
\title{Estimate correlations among Google Trend keywords}
\usage{
trend_correlation(tbl)
}
\arguments{
\item{tbl}{Data frame output from Google Trends API}
}
\value{
Data frame of correlation data
}
\description{
Estimates the correlations across keywords with a correlation table.
}
\examples{
\dontrun{
trend_correlation(tbl)}
}
